feat(agent): introduce specific DIDComm controller/endpoints classes
