+++ shared/core/src/main/scala/org/hyperledger/identus/shared/models/Failure.scala
@@ -1,6 +1,7 @@
 package org.hyperledger.identus.shared.models
 
 trait Failure {
+  val namespace: String
   val statusCode: StatusCode
   val userFacingMessage: String
 }
@@ -10,6 +11,7 @@ sealed class StatusCode(val code: Int)
 object StatusCode {
   val BadRequest: StatusCode = StatusCode(400)
   val NotFound: StatusCode = StatusCode(404)
+  val UnprocessableContent: StatusCode = StatusCode(422)
 
   val InternalServerError: StatusCode = StatusCode(500)
 }
